containers/pkgsite: Add

+++ script/vendor-pkgsite.sh
@@ -6,25 +6,24 @@ source "script/lib/vendor.sh"
 OWNER="golang"
 REPO="pkgsite"
 IMPORTPATH="golang.org/x/pkgsite"
-COMMIT="5ae7386a8ff7d9caec4c44f8ce68f402793f3954"
+COMMIT="aa5dca7496c53df88ed8bb5279c4d76c02c78bd3"
 
-PATCH_DIR=$(cd $(dirname $0)/${REPO}; pwd)
-THIRD_PARTY_DIR=$(cd $(dirname $0)/../third_party; pwd)
+if [ -z "$BUILD_WORKSPACE_DIRECTORY" ]; then
+    echo "Please exec via bazel"
+    echo "bazel run //script:vendor_pkgsite"
+    exit 1
+fi
+
+PATCH_DIR="${BUILD_WORKSPACE_DIRECTORY}/script/pkgsite"
+THIRD_PARTY_DIR="${BUILD_WORKSPACE_DIRECTORY}/third_party"
 TARGET_DIR="${THIRD_PARTY_DIR}/${REPO}"
-
 if [ -d "${TARGET_DIR}" ]; then
   rm -rf "${TARGET_DIR}"
 fi
-mkdir -p "${TARGET_DIR}"
 
 download_repository_from_github "${TARGET_DIR}" "${OWNER}" "${REPO}" "${COMMIT}"
 
-remove_git_files "${TARGET_DIR}"
-remove_bazel_files "${TARGET_DIR}"
-remove_unnecessary_go_files "${TARGET_DIR}"
-
-cd "${TARGET_DIR}"
-echo $COMMIT > COMMIT
+remove_unnecessary_files "${TARGET_DIR}"
 
 generate_build_file "${TARGET_DIR}" "${IMPORTPATH}"
